requirements.txt
anyio==3.6.2
bleach==6.0.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-log==0.4.0
colorama==0.4.6
cryptography==40.0.1
deprecation==2.1.0
docutils==0.19
dotty-dict==1.3.1
fastapi==0.92.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
gotrue==1.0.1
h11==0.14.0
honeybadger==0.15.1
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
invoke==1.7.3
jaraco.classes==3.2.3
jeepney==0.8.0
keyring==23.13.1
more-itertools==9.1.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pkginfo==1.9.6
postgrest==0.10.6
protobuf==4.22.0
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.5
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==3.13.0
python-semantic-release==7.33.2
pytz==2022.7.1
readme-renderer==37.3
realtime==1.0.0
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==1.5.0
rsa==4.9
SecretStorage==3.3.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.25.0
storage3==0.5.2
StrEnum==0.4.10
supabase==1.0.3
supafunc==0.2.2
tomlkit==0.11.7
tqdm==4.65.0
twine==3.8.0
typing-extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
webencodings==0.5.1
websockets==10.4
zipp==3.15.0
